refactor/move goldie test to product_test.go
