+++ app/server/api_test/main_test.go
@@ -1,15 +1,18 @@
 package api_test
 
 import (
+	"bytes"
+	"encoding/json"
+	"testing"
+
+	"github.com/gin-gonic/gin"
+	"gopkg.in/testfixtures.v2"
+
 	"github/code-kakitai/code-kakitai/infrastructure/mysql/db"
 	dbTest "github/code-kakitai/code-kakitai/infrastructure/mysql/db/db_test"
 	"github/code-kakitai/code-kakitai/infrastructure/mysql/db/dbgen"
 	"github/code-kakitai/code-kakitai/presentation/settings"
 	"github/code-kakitai/code-kakitai/server/route"
-	"testing"
-
-	"github.com/gin-gonic/gin"
-	"gopkg.in/testfixtures.v2"
 )
 
 var (
@@ -56,4 +59,16 @@ func resetTestData(t *testing.T) {
 	if err := fixtures.Load(); err != nil {
 		t.Fatal(err)
 	}
+}
+
+// Jsonのフォーマットを整える
+func formatJSON(t *testing.T, b []byte) []byte {
+	t.Helper()
+
+	var out bytes.Buffer
+	err := json.Indent(&out, b, "", "  ")
+	if err != nil {
+		t.Fatal(err)
+	}
+	return out.Bytes()
 }
